@@ -1,13 +1,14 @@
 {
   "name": "detect-libc",
-  "version": "1.0.3",
+  "version": "2.0.1",
   "description": "Node.js module to detect the C standard library (libc) implementation family and version",
   "main": "lib/detect-libc.js",
-  "bin": {
-    "detect-libc": "./bin/detect-libc.js"
-  },
+  "files": [
+    "lib/",
+    "index.d.ts"
+  ],
   "scripts": {
-    "test": "semistandard && nyc --reporter=lcov ava"
+    "test": "semistandard && nyc --reporter=lcov --check-coverage --branches=100 ava test/unit.js"
   },
   "repository": {
     "type": "git",
@@ -24,12 +25,12 @@
   ],
   "license": "Apache-2.0",
   "devDependencies": {
-    "ava": "^0.23.0",
-    "nyc": "^11.3.0",
-    "proxyquire": "^1.8.0",
-    "semistandard": "^11.0.0"
+    "ava": "^2.4.0",
+    "nyc": "^15.1.0",
+    "proxyquire": "^2.1.3",
+    "semistandard": "^14.2.3"
   },
   "engines": {
-    "node": ">=0.10"
+    "node": ">=8"
   }
 }
