@@ -10,20 +10,53 @@ import { GoogleMapsModule } from '@angular/google-maps';
 // import { Geolocation } from '@ionic-native/geolocation/ngx';
 import { Geolocation } from '@awesome-cordova-plugins/geolocation/ngx';
 import { NativeGeocoder } from '@ionic-native/native-geocoder/ngx';
+import { AuthApiService } from './service/api.authService';
+
+import { 
+  GoogleLoginProvider,
+  FacebookLoginProvider,
+  SocialLoginModule,
+  SocialAuthServiceConfig
+ } from 'angularx-social-login';
+import { AuthModule } from './auth/auth.module';
 
 @NgModule({
   declarations: [AppComponent],
   entryComponents: [],
-// <<<<<<< jiam
-  imports: [BrowserModule, IonicModule.forRoot(), AppRoutingModule, GoogleMapsModule, HttpClientModule],
-  providers: [Geolocation, NativeGeocoder, { provide: RouteReuseStrategy, useClass: IonicRouteStrategy }],
-// =======
-//   imports: [BrowserModule, 
-//           IonicModule.forRoot(),  
-//           AppRoutingModule,
-//           HttpClientModule],
-//   providers: [{ provide: RouteReuseStrategy, useClass: IonicRouteStrategy }],  
-// >>>>>>> main
+  imports: [
+    BrowserModule, 
+    IonicModule.forRoot(), 
+    AppRoutingModule, 
+    GoogleMapsModule, 
+    HttpClientModule,
+    SocialLoginModule,
+    AuthModule
+  ],
+  providers: [
+    Geolocation, 
+    NativeGeocoder, 
+    AuthApiService,
+    { 
+      provide: RouteReuseStrategy, 
+      useClass: IonicRouteStrategy 
+    }, 
+    {
+      provide: 'SocialAuthServiceConfig',
+      useValue: {
+        autoLogin: false,
+        providers: [
+          {
+            id: GoogleLoginProvider.PROVIDER_ID,
+            provider: new GoogleLoginProvider('364897496119-rrteeo2ddr72f6j3je7nj5t39b7uc7ve.apps.googleusercontent.com')
+          },
+          {
+            id: FacebookLoginProvider.PROVIDER_ID,
+            provider: new FacebookLoginProvider('1008674980068196')
+          }
+        ]
+      } as SocialAuthServiceConfig,
+    }
+  ],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
