@@ -5,48 +5,61 @@ const bcrypt = require('bcrypt')
 //salt round represents how many salt 
 const saltRounds = 10
 const jwt = require('jsonwebtoken')
+const { async } = require('rxjs')
 
 const userSchema = mongoose.Schema({
-    firstname: {
+    username: {
         type: String,
-        maxLength: 50,
-        required: true
-    },
-    lastname: {
-        type: String,
-        maxLength: 20,
-        required: true
+        maxLength: 30,
+        // required: true
     },
     email: {
         type: String,
         trim: true, //removes all the whitespaces
         unique: 1,
-        required: true
+        // required: true
     },
     password: {
         type: String,
-        minLength: 10,
+        minLength: 1,
     }, 
     workLevel: {
         type: String,
         enum: ['Secondary School', 'Junior College', 
-                'Polytechnic', 'University', 'Professional Work']
+                'Polytechnic', 'University'],
+        // required: true
+        // edited secondary school to secondary, removed working
     },
     bio: {
         type: String,
         maxlength: 200
     },
-    token: {
-        type: String
-    },
-    tokenExp: {
-        type: Number
-    },
     provider: {
         type: String,
-        enum: ['google', 'local']
+        enum: ['GOOGLE', 'LOCAL', 'FACEBOOK'],
+        // required: true
+    },
+    telegram: {
+        type: String,
+        default: null
     },
     googleid: {
+        type: String,
+        default: null
+    },
+    facebookid : {
+        type: String,
+        default: null
+    },
+    photoURL: {
+        type: String,
+        default: null
+    },
+    photoKey: {
+        type: String,
+        default: null
+    },
+    token: {
         type: String
     }
     
@@ -77,15 +90,31 @@ userSchema.pre('save', function(next){
     }
 })
 
-userSchema.methods.comparePassword = function(plainPassword, cb){
+userSchema.methods.comparePassword = async function(plainPassword, cb){
     //plainpassword is the password entered by the user
     //we cannot decrypt the encrypted password to compare with the plain password
     //so we encrypt the plainpassword then compare it with the encrypted string
-    bcrypt.compare(plainPassword, this.password, (err, isMatch) => {
-        if (err) return cb(err),
-            cb(null, isMatch) //if password match, return null as error and return isMatch = True
+    console.log('within user method')
+    console.log(plainPassword)
+    
+    console.log(this.password)
+    // await bcrypt.compare(plainPassword, this.password, function (err, isMatch)  {
+    //     //console.log(isMatch)
+    //     console.log('Comparing Password')
+    //     console.log(isMatch)
+    //     if (err) return cb(err),
+    //         cb(null, isMatch)
+        
+    //     //if password match, return null as error and return isMatch = True
 
-    })
+    // })
+    
+    try {
+        return await bcrypt.compare(plainPassword, this.password)
+    } catch (error) {
+        console.log(error)
+    }
+    return false
 }
 
 userSchema.methods.generateToken = function(callback){
