@@ -2,15 +2,17 @@ const express = require('express');
 const mainRouter = express.Router();
 const mongoose = require('mongoose')
 const config = require('../config/key')
+const session = require('express-session')
 
-mainRouter.use('/api/localauth', require('./userRoutes/localAuthController'))
-mainRouter.use('/api/jio', require('./jioRoutes/jioController'))
-mainRouter.use('/api/google', require('./userRoutes/googleController'))
-
-
+mainRouter.use('/localauth', require('./userRoutes/localAuthController'))
+mainRouter.use('/jio', require('./jioRoutes/jioController'))
+mainRouter.use('/google', require('./userRoutes/googleController'))
+mainRouter.use('/profile', require('./userRoutes/profileController'))
 
 mainRouter.get('/test', (req, res) => {
     res.send('Test dumb')
 })
 
 module.exports = mainRouter;
+
+
