@@ -8,75 +8,112 @@ const config = require('../../config/key')
 const { response, application } = require('express')
 const cookieparser = require('cookie-parser')
 const { auth } = require('../../middleware/auth')
+const { Jio } = require('../../models/Jio')
+const passport = require('./passport')
+const Passport = require('passport')
+
+
+//const imageController = require('./imageController')
 
 
 //application/x-www-form-urlencoded 
 router.use(bodyParser.urlencoded({extended: true}))
 router.use(bodyParser.json())
 router.use(cookieparser())
-
-// mongoose.connect(
-//     config.mongoURI, {})
-//     .then(() => console.log('MongoDB connected'))
-//     .catch(err => console.log(err))
-
-
+router.use(express.json())
+//router.use(express.bodyParser())
 
 router.get('/', (req, res) => res.send('stfu bitch'))
  
 //The following function is mounted on the /register path. It is executed for any type
 //of HTTP request on the /register path
 
-router.post('/local-signup', (req, res) => {
+router.post('/register', (req, res) => {
     //회원가입할때 필요한 정보들을 client에서 가져오면 그것들을 데이터베이스에 넣어준다
+    console.log('Backend: Local Register')
     const user = new User(req.body)
     //crypting password before saving to the database
+    User.findOne({email: req.body.email}, async (err, user) => {
+        if (user){
+            return res.json({
+                result: false
+            })
+        }
+    })
     user.save((err, doc) => {
         console.log(doc)
         if (err) return res.json({ success: false, err})
         return res.status(200).json({
-            success: true
+            success: true,
+            userID: user._id
         })
     })
-
 })
 
 
 router.post('/local-login', (req, res) => {
-    //first find whether requested email address exists in the database
-    //mongodb method
-    User.findOne({email: req.body.email}, (err, user) => {
+
+    User.findOne({email: req.body.email}, async (err, user) => {
         if(!user){ //if user does not exist
             return res.json({
                 loginSuccess: false,
                 message: "Email address does not exist"
             })
         } 
-    
-    //if email address exists in the database check whether password is correct
-        user.comparePassword(req.body.password, (err, isMatch) => {
-            if(!isMatch){
-                return res.json({
-                    loginSuccess: false,
-                    message: "Incorrect Password"
-                })
-            } 
-        }) 
-    //if password correct, create a token for the user
-        user.generateToken((err, user) => {
-            if (err) return res.status(400).send(err) //400 means error
-            
-            // user contains token. we need to save it somewhere (cookie? local storage?)
-            //to save it to cookie we need cookieparser
-            res.cookie("x_auth", user.token)
-            .status(200) //means success
-            .json({
-                loginSuccess: true,
-                userID: user._id
+
+        const validated = await user.comparePassword(req.body.password)
+        if (!validated) {
+            return res.json({
+                loginSuccess: false,
+                message: "incorrect password"
             })
+        } else {
+            console.log('validated')
+            req.session.user = user
+            console.log(req.session.user)
+            req.session.save((err) => {
+                if (err) {
+                    return res.json({
+                        loginSuccess: false,
+                        error: err
+                    })
+                } else {
+                    return res.json({
+                        loginSuccess: true,
+                        userID: user._id
+                    })
+                }
+            })
+        }
+    })
+})
+
+
+
+
+// Verify if account exists via email matching
+router.post('/verify-account', (req, res) => {
+    User.findOne({email:req.body.email}, async (err, user) => {
+        console.log(req.body.email + "verify account");
+        if(!user){
+            return res.json({
+                verifyresult: false
+            })
+        }
+        req.session.user = user
+        req.session.save()
+        return res.json({
+            userID: user._id,
+            result: true
         })
     })
 })
+
+
+
+
+
+
 
 
 router.get('/auth', auth, (req, res) => {  //middleware implementation
@@ -112,6 +149,7 @@ router.get('/local-logout', auth, (req, res) => {
 
     //just removing token
 })
+
 
 
 
