@@ -15,10 +15,6 @@ const routes: Routes = [
     loadChildren: () => import('./settings/settings.module').then( m => m.SettingsPageModule)
   },
   {
-    path: 'chat-indiv',
-    loadChildren: () => import('./chat-indiv/chat-indiv.module').then( m => m.ChatIndivPageModule)
-  },
-  {
     path: 'login',
     loadChildren: () => import('./login/login.module').then( m => m.LoginPageModule)
   },
@@ -47,9 +43,38 @@ const routes: Routes = [
     loadChildren: () => import('./external-register/external-register.module').then( m => m.ExternalRegisterPageModule)
   },
   {
-    path: 'spot-details/:place_id',
-    loadChildren: () => import('./spot-details/spot-details.module').then((m) => m.SpotDetailsPageModule),
+    path: 'jio-details',
+    loadChildren: () => import('./jio-details/jio-details.module').then( m => m.JioDetailsPageModule)
   },
+  {
+    path: 'jio-list',
+    loadChildren: () => import('./jio-list/jio-list.module').then( m => m.JioListPageModule)
+  },
+  {
+    path: 'jio-form',
+    loadChildren: () => import('./jio-form/jio-form.module').then( m => m.JioFormPageModule)
+  },
+  {
+    path: 'jio-details-accept',
+    loadChildren: () => import('./jio-details-accept/jio-details-accept.module').then( m => m.JioDetailsAcceptPageModule)
+  },
+  {
+    path: 'jio-history',
+    loadChildren: () => import('./jio-history/jio-history.module').then( m => m.JioHistoryPageModule)
+  },
+  {
+    path: 'jio-details-pending',
+    loadChildren: () => import('./jio-details-pending/jio-details-pending.module').then( m => m.JioDetailsPendingPageModule)
+  },
+  {
+    path: 'create-jio',
+    loadChildren: () => import('./create-jio/create-jio.module').then( m => m.CreateJioPageModule)
+  }
+  // {
+  //   path: 'jio-create',
+  //   loadChildren: () => import('./jio-create/jio-create.module').then( m => m.JioCreatePageModule)
+  // }
+
 ];
 @NgModule({
   imports: [
