@@ -4,6 +4,7 @@ import { defaultMaxListeners } from 'events';
 import { fromEventPattern } from 'rxjs';
 import { GooglePlacesAPIService } from '../service/google-places-api.service';
 import { Geolocation } from '@awesome-cordova-plugins/geolocation/ngx';
+import { LocationDetailsPageRoutingModule } from '../location-details/location-details-routing.module';
 
 declare var google: any;
 
@@ -13,74 +14,93 @@ declare var google: any;
   styleUrls: ['map.page.scss']
 })
 export class MapPage {
-  public radius:any;
-  map:any;
-  
-  @ViewChild('map', {read: ElementRef, static: false}) mapRef: ElementRef;
+  public radius: number = 2000;
+  map: any;
 
-  @ViewChild('search', {static: false}) search: IonSearchbar;
+  @ViewChild('map', { read: ElementRef, static: false }) mapRef: ElementRef;
+
+  @ViewChild('search', { static: false }) search: IonSearchbar;
 
   infoWindows: any = [];
   public markers: any = []
-
   private currentLoc: any = []
+
+  private sg_lat = 1.3521
+  private sg_lng = 103.8198
 
   // constructor(public api:GooglePlacesAPIService){}
 
-  constructor(public api:GooglePlacesAPIService, private geolocation: Geolocation) {}
-  
-  ngOnInit(){}
+  constructor(public api: GooglePlacesAPIService, private geolocation: Geolocation) { }
 
-
-  getCurrentLocation(){
-    this.geolocation.getCurrentPosition().then((resp) => {
-      let current:object;
-      current = {
-        latitude:resp.coords.latitude,
-        longitude:resp.coords.longitude
-      }
-      this.currentLoc.push(current);    
-      console.log(current);
-      this.addMarkersToMap(this.currentLoc);
-     }).catch((error) => {
-       console.log('Error getting location', error);
-     });
+  ngOnInit() {
   }
 
-  updateSearchResults(){
+  getCurrentLocation() {
     console.log("clicked")
+    // this.getLocationsNearUser();
+    this.addCurrentLocToMap(this.currentLoc);
+
+    this.map.panTo({ lat: this.currentLoc[0].latitude, lng: this.currentLoc[0].longitude });
+
+    console.log("getting locations near user")
+    this.getLocationsNearUser();
+    // setTimeout(() => {
+    //   console.log("getting locations near user")
+    //   this.getLocationsNearUser();
+    // },5000); 
+
+    setTimeout(() => {
+      console.log("setting zoom")
+      this.map.setZoom(15)
+    },1000); 
+    
+
+    
+  }
+
+  async getLocationsNearUser() {
     let loc = String(this.currentLoc[0].latitude) + '%2C' + String(this.currentLoc[0].longitude);
-    this.api.getLocations(this.radius, loc).subscribe(result=>{
+    this.api.getLocations(this.radius, loc).subscribe(result => {
       let locationData = result;
-      let newLocation:object;
-      for (let key in locationData){
+      let newLocation: object;
+      for (let key in locationData) {
         newLocation = {
           title: locationData[key].name,
           latitude: String(locationData[key].geometry.location.lat),
           longitude: String(locationData[key].geometry.location.lng),
           rating: locationData[key].rating,
-          price: "TBC"
+          // src: this.fetchPhoto(locationData[key].photos[0].photo_reference, 400)
         }
+
         this.markers.push(newLocation);
       }
     })
-
-    const location = new google.maps.LatLng(this.currentLoc.latitude , this.currentLoc.longitude);
-    // const options = {
-    //   center: location,
-    //   zoom: 12,
-    //   disableDefaultUI: true,
-    //   keyboardShortcuts: false
-    // }
-    // this.map = new google.maps.Map(this.mapRef.nativeElement, options)
-    this.addMarkersToMap(this.markers);
-    console.log("clicked")
+    console.log("trying markers to map");
+    setTimeout(()=>{
+      this.addMarkersToMap(this.markers)
+    },1000);
+    console.log("supposed to add markers to map")
   }
 
-  ionViewDidEnter(){
-    this.showMap();
-    this.getCurrentLocation();
+  ionViewDidEnter() {
+    this.showMap(this.sg_lat, this.sg_lng);
+
+    this.geolocation.getCurrentPosition().then((resp) => {
+      this.currentLoc = [{
+        latitude: resp.coords.latitude,
+        longitude: resp.coords.longitude
+      }];
+    }).catch((error) => {
+      console.log('Error getting location', error);
+    });
+
+    
+    // this.getCurrentLocation();
+    // this.addMarkersToMap(this.currentLoc);
+    // this.showMap(this.currentLoc.latitude, this.currentLoc.longitude);
+
   }
+
 
   // addMarkerToMap(marker){
   //   let position = new google.maps.LatLng(marker.latitude, marker.longitude);
@@ -89,8 +109,9 @@ export class MapPage {
   //   console.log(marker + "within add function")
   // }
 
-  addMarkersToMap(markers){
-    for(let marker of markers){
+  addMarkersToMap(markers) {
+    console.log("inside addmarkerstomap function (before for loop)")
+    for (let marker of markers) {
       let position = new google.maps.LatLng(marker.latitude, marker.longitude);
       let mapMarker = new google.maps.Marker({
         position: position,
@@ -98,175 +119,82 @@ export class MapPage {
         latitude: marker.latitude,
         longitude: marker.longitude,
         rating: marker.rating,
-        price: marker.price
+        price: marker.price,
+      });
+      console.log("inside addmarkerstomap function")
+      mapMarker.setMap(this.map);
+      this.addInfoWindowToMarker(mapMarker);
+    }
+  }
+
+  addCurrentLocToMap(markers) {
+    for (let marker of markers) {
+      let position = new google.maps.LatLng(marker.latitude, marker.longitude);
+      let mapMarker = new google.maps.Marker({
+        position: position,
+        title: marker.title,
+        latitude: marker.latitude,
+        longitude: marker.longitude,
+        rating: marker.rating,
+        price: marker.price,
+        icon: 'http://maps.google.com/mapfiles/ms/icons/green-dot.png'
       });
 
       mapMarker.setMap(this.map);
-      this.addInfoWindowToMarker(mapMarker);
+      // this.addInfoWindowToMarker(mapMarker);
     }
   }
 
   addInfoWindowToMarker(marker){
     let infoWindowContent = '<div id="content" style="color:black">' + 
                               '<h2 id="firstHeading" class="firstHeading">' + marker.title + '</h2>' +
-                              '<p>Rating: ' + marker.rating + '/5</p>' +
-                              '<p>Price: ' + marker.price+ '</p>' +
+                              '<h4> Rating: ' + marker.rating + '</h4>' +
                               '<ion-button id="navigate">Navigate</ion-button>' +
                             '</div>';
 
-    // let infoWindowContent = 'hey hey';                     
 
     let infoWindow = new google.maps.InfoWindow({
       content: infoWindowContent
     })
 
     marker.addListener('click', () => {
-        this.closeAllInfoWindows();
-        infoWindow.open(this.map, marker)
+      this.closeAllInfoWindows();
+      infoWindow.open(this.map, marker)
 
-        google.maps.event.addListenerOnce(infoWindow, 'domready', ()=> {
-          document.getElementById('navigate').addEventListener('click', ()=>{
-            console.log('button clicked!');
-            window.open('https://www.google.com/maps/dir/?api=1&origin=' + this.currentLoc.latitude + ',' 
-            + this.currentLoc.longitude + '&destination=' + marker.latitude + ',' + marker.longitude);
-            // window.open('https://maps.google.com/maps?saddr=Current+Location&daddr=' + marker.latitude + ',' + marker.longitude);
-            // window.open('https://www.google.com/maps/dir/?api1=&destination=' + marker.latitude + ',' + marker.longitude);
-          });
+      google.maps.event.addListenerOnce(infoWindow, 'domready', () => {
+        document.getElementById('navigate').addEventListener('click', () => {
+          console.log('button clicked!');
+          console.log(this.currentLoc[0].latitude);
+          window.open('https://www.google.com/maps/dir/?api=1&origin=' + this.currentLoc[0].latitude + ','
+            + this.currentLoc[0].longitude + '&destination=' + marker.latitude + ',' + marker.longitude);
+        });
 
-        })
+      })
     });
     this.infoWindows.push(infoWindow);
   }
 
-  closeAllInfoWindows(){
-    for(let window of this.infoWindows){
+  closeAllInfoWindows() {
+    for (let window of this.infoWindows) {
       window.close();
     }
   }
 
-  showMap(){
-    const location = new google.maps.LatLng(1.3521, 103.8198);
+  showMap(ref_lat, ref_lng) {
+    const location = new google.maps.LatLng(ref_lat, ref_lng);
     const options = {
       center: location,
       zoom: 12,
       disableDefaultUI: true,
       keyboardShortcuts: false,
-      // styles: [
-      //   { elementType: "geometry", stylers: [{ color: "#1b2430" }] },
-      //   // 242f3e
-      //   { elementType: "labels.text.stroke", stylers: [{ color: "#242f3e" }] },
-      //   { elementType: "labels.text.fill", stylers: [{ color: "#746855" }] },
-      //   {
-      //     featureType: "administrative.locality",
-      //     elementType: "labels.text.fill",
-      //     stylers: [{ color: "#d59563" }],
-      //   },
-      //   {
-      //     featureType: "poi",
-      //     elementType: "labels.text.fill",
-      //     stylers: [{ color: "#d59563" }],
-      //   },
-      //   {
-      //     featureType: "poi.park",
-      //     elementType: "geometry",
-      //     stylers: [{ color: "#263c3f" }],
-      //   },
-      //   {
-      //     featureType: "poi.park",
-      //     elementType: "labels.text.fill",
-      //     stylers: [{ color: "#6b9a76" }],
-      //   },
-      //   {
-      //     featureType: "road",
-      //     elementType: "geometry",
-      //     stylers: [{ color: "#38414e" }],
-      //   },
-      //   {
-      //     featureType: "road",
-      //     elementType: "geometry.stroke",
-      //     stylers: [{ color: "#212a37" }],
-      //   },
-      //   {
-      //     featureType: "road",
-      //     elementType: "labels.text.fill",
-      //     stylers: [{ color: "#9ca5b3" }],
-      //   },
-      //   {
-      //     featureType: "road.highway",
-      //     elementType: "geometry",
-      //     stylers: [{ color: "#746855" }],
-      //   },
-      //   {
-      //     featureType: "road.highway",
-      //     elementType: "geometry.stroke",
-      //     stylers: [{ color: "#1f2835" }],
-      //   },
-      //   {
-      //     featureType: "road.highway",
-      //     elementType: "labels.text.fill",
-      //     stylers: [{ color: "#f3d19c" }],
-      //   },
-      //   {
-      //     featureType: "transit",
-      //     elementType: "geometry",
-      //     stylers: [{ color: "#2f3948" }],
-      //   },
-      //   {
-      //     featureType: "transit.station",
-      //     elementType: "labels.text.fill",
-      //     stylers: [{ color: "#d59563" }],
-      //   },
-      //   {
-      //     featureType: "water",
-      //     elementType: "geometry",
-      //     stylers: [{ color: "#17263c" }],
-      //   },
-      //   {
-      //     featureType: "water",
-      //     elementType: "labels.text.fill",
-      //     stylers: [{ color: "#515c6d" }],
-      //   },
-      //   {
-      //     featureType: "water",
-      //     elementType: "labels.text.stroke",
-      //     stylers: [{ color: "#17263c" }]
-      //   }]
-      
     }
 
     this.map = new google.maps.Map(this.mapRef.nativeElement, options)
-    this.addMarkersToMap(this.markers);
+    // this.addMarkersToMap(this.markers);
   }
 
-  ionChange(event){
+  ionChange(event) {
     console.log(event.detail.value);
   }
-
-  // initMap() {
-  //   navigator.geolocation.getCurrentPosition((location) => {
-  //     console.log(location);
-  //     map = new google.maps.Map(this.mapElement.nativeElement, {
-  //       center: {lat: location.coords.latitude, lng: location.coords.longitude},
-  //       zoom: 15
-  //     });
-  
-  //     infowindow = new google.maps.InfoWindow();
-  //     var service = new google.maps.places.PlacesService(map);
-  //     service.nearbySearch({
-  //       location: {lat: location.coords.latitude, lng: location.coords.longitude},
-  //       radius: 1000,
-  //       type: ['store']
-  //     }, (results,status) => {
-  //       if (status === google.maps.places.PlacesServiceStatus.OK) {
-  //         for (var i = 0; i < results.length; i++) {
-  //           this.createMarker(results[i]);
-  //         }
-  //       }
-  //     });
-  //   }, (error) => {
-  //     console.log(error);
-  //   }, options);
-  //   var myplace = {lat: -33.8665, lng: 151.1956};
-  // }
 
 }
