@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { GooglePlacesAPIService } from '../service/google-places-api.service';
 import { SpotPage } from '../spot/spot.page';
+import { Geolocation } from '@awesome-cordova-plugins/geolocation/ngx';
 
 @Component({
   selector: 'app-spot-details',
@@ -10,21 +11,52 @@ import { SpotPage } from '../spot/spot.page';
 export class SpotDetailsPage implements OnInit {
 
   place_id: string;
+  linearDistance: string;
+  src: string;
+  lat: string;
+  lng: string;
   spotDetails: any;
+  currentLoc: any = [];
 
-  constructor(public api:GooglePlacesAPIService) { }
+  constructor(public api:GooglePlacesAPIService, private geolocation: Geolocation) {
+    // this.getCurrentLocation()
+   }
+
+  getCurrentLocation() {
+    this.geolocation.getCurrentPosition().then((resp) => {
+      this.currentLoc = [{
+        latitude:resp.coords.latitude,
+        longitude:resp.coords.longitude
+      }];    
+     }).catch((error) => {
+       console.log('Error getting location', error);
+     });
+
+  }
 
   ngOnInit() {
     this.place_id = sessionStorage.getItem("place_id")
+    this.linearDistance = sessionStorage.getItem('linearDistance');
+    this.src = sessionStorage.getItem('src');
+    this.lat = sessionStorage.getItem('lat');
+    this.lng = sessionStorage.getItem('lng');
     this.fetchLocation(this.place_id);
+    this.getCurrentLocation();
   }
 
   fetchLocation(place_id:string){
     this.api.getPlaceDetails(place_id).subscribe(result => {
       console.log(result);
       this.spotDetails = Array.of(result);
-      this.spotDetails.src = "https://static.vecteezy.com/system/resources/thumbnails/000/599/173/small/coffee.jpg";
+      this.spotDetails[0].result.linearDistance = this.linearDistance;
+      this.spotDetails[0].result.src = this.src;
     })
+  }
+
+  navigateToPlace(){
+    // console.log("currentLoc " + this.currentLoc[0])
+    window.open('https://www.google.com/maps/dir/?api=1&origin=' + this.currentLoc[0].latitude + ',' 
+            + this.currentLoc[0].longitude + '&destination=' + this.lat + ',' + this.lng);
   }
 
 }
