@@ -5,12 +5,14 @@ const {
     getDistance, 
     getPlaceSearchID, 
     getPlaceCoordinates,
-    getPlaceDetails } = require('../controllers/places');
+    getPlaceDetails,
+    getPlacePhoto } = require('../controllers/places');
 
 router.route('/search').get(getLocations);
 router.route('/distance').get(getDistance);
 router.route('/placesearch').get(getPlaceSearchID);
 router.route('/placecoordinates').get(getPlaceCoordinates);
 router.route('/placedetails').get(getPlaceDetails);
+router.route('/placephoto').get(getPlacePhoto);
 
 module.exports = router;
